requirements.txt
streamlit ==  1.37.1
pillow == 10.4.0
